test(hygiene): split ai smoke guardrails
Separate runtime smoke from static text and file guardrails so the
hygiene suite groups checks by purpose instead of one catch-all file.

Refs #2802
